+++ package.json
@@ -1,40 +1,55 @@
 {
-	"name": "@antv/expr",
-	"version": "1.0.1",
-	"description": "A secure, high-performance expression evaluator for dynamic chart rendering",
-	"main": "dist/index.cjs.js",
-	"module": "dist/index.esm.js",
-	"types": "dist/index.d.ts",
-	"files": ["dist", "LICENSE", "README.md", "package.json"],
-	"scripts": {
-		"build": "rollup -c",
-		"test": "vitest run --coverage",
-		"benchmark": "vitest bench",
-		"prepublishOnly": "pnpm run test && pnpm run build"
-	},
-	"keywords": [
-		"expression",
-		"evaluator",
-		"parser",
-		"secure",
-		"antv",
-		"chart",
-		"expr"
-	],
-	"repository": {
-		"type": "git",
-		"url": "https://github.com/antvis/expr"
-	},
-	"license": "MIT",
-	"devDependencies": {
-		"@biomejs/biome": "1.9.4",
-		"@rollup/plugin-node-resolve": "^16.0.0",
-		"@rollup/plugin-terser": "^0.4.4",
-		"@rollup/plugin-typescript": "^12.1.2",
-		"@vitest/coverage-v8": "^3.0.8",
-		"expr-eval": "^2.0.2",
-		"rollup": "^4.34.6",
-		"tslib": "^2.8.1",
-		"vitest": "^3.0.8"
-	}
+  "name": "@antv/expr",
+  "version": "1.0.1",
+  "description": "A secure, high-performance expression evaluator for dynamic chart rendering",
+  "main": "dist/index.cjs.js",
+  "module": "dist/index.esm.js",
+  "types": "dist/index.d.ts",
+  "files": [
+    "dist",
+    "LICENSE",
+    "README.md",
+    "package.json"
+  ],
+  "scripts": {
+    "build": "rollup -c && npm run size",
+    "test": "vitest run --coverage",
+    "size": "limit-size",
+    "benchmark": "vitest bench",
+    "prepublishOnly": "pnpm run test && pnpm run build"
+  },
+  "keywords": [
+    "expression",
+    "evaluator",
+    "parser",
+    "secure",
+    "antv",
+    "chart",
+    "expr"
+  ],
+  "devDependencies": {
+    "@biomejs/biome": "1.9.4",
+    "@rollup/plugin-node-resolve": "^16.0.0",
+    "@rollup/plugin-terser": "^0.4.4",
+    "@rollup/plugin-typescript": "^12.1.2",
+    "@vitest/coverage-v8": "^3.0.8",
+    "expr-eval": "^2.0.2",
+    "limit-size": "^0.1.4",
+    "rollup": "^4.34.6",
+    "tslib": "^2.8.1",
+    "vitest": "^3.0.8"
+  },
+  "limit-size": [{
+    "path": "dist/index.cjs.js",
+    "limit": "8 Kb"
+  }, {
+    "path": "dist/index.cjs.js",
+    "limit": "3 Kb",
+    "gzip": true
+  }],
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/antvis/expr"
+  },
+  "license": "MIT"
 }
